Move common.exceptions to common.http.exceptions
