+++ starlette_web/auth/views.py
@@ -12,7 +12,7 @@ from starlette_web.common.statuses import ResponseStatus
 from starlette_web.core import settings
 from starlette_web.common.views import BaseHTTPEndpoint
 from starlette_web.common.utils import send_email, get_logger
-from starlette_web.common.exceptions import AuthenticationFailedError, InvalidParameterError
+from starlette_web.common.http.exceptions import AuthenticationFailedError, InvalidParameterError
 from starlette_web.auth.models import User, UserSession, UserInvite
 from starlette_web.auth.hasher import PBKDF2PasswordHasher, get_salt
 from starlette_web.auth.backend import AdminRequiredAuthBackend, LoginRequiredAuthBackend
